Adding all the present RetroRemissions Functional Tests....
